stripe_api/requirements.txt
stripe>=4.0.0
requests==2.28.2
django==3.2.18
djangorestframework==3.14.0
python-dotenv==0.21.1
gunicorn==20.0.4
psycopg2-binary==2.8.6
asgiref>=3.3.2
sqlparse==0.3.1
pytz==2020.1